backend/requirements.txt
# Core FastAPI and Server Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database Dependencies
singlestoredb==1.2.0
pymysql==1.1.0
sqlalchemy==2.0.23

# Data Processing and Analysis
pandas==2.1.3
numpy==1.25.2
plotly==5.17.0

# Dash Application Dependencies
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Requests and APIs
requests==2.31.0
httpx==0.25.2

# AI and Machine Learning (Optional - uncomment if using AI features)
# openai==1.3.0
# scikit-learn==1.3.2
# yfinance==0.2.28

# Testing Dependencies (Optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1

# Development Dependencies (Optional)
# black==23.11.0
# flake8==6.1.0
# mypy==1.7.1

# Additional Utilities
tenacity==8.2.3
python-dateutil==2.8.2
pytz==2023.3 